api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
google-auth==2.49.1
google-auth-oauthlib==1.3.1
gotrue==2.12.4
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==0.2.43
langchain-text-splitters==0.2.1
langsmith==0.1.147
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy>=1.26.0
oauthlib==3.3.1
openai==1.35.0
orjson==3.11.8
packaging==24.2
pillow==12.2.0
postgrest==0.16.11
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.20.0
PyJWT==2.12.1
PyMuPDF==1.27.2.2
pypdf==4.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.3
realtime==1.0.6
regex==2026.3.32
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.7
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
